requirements.txt
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0

